January/20th-Jan-Tues/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "pymysql>=1.1.2",
    "python-decouple>=3.8",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
